\name{getGlobalUrl}
\alias{getGlobalUrl}
\title{Get the global URL for which DAPI server is used.}
\description{Get the global URL for which DAPI server is used. Uses the form: http://localhost:8080/DAPI}
\usage{
getGlobalUrl()
}
\value{Returns results from get call -- a string URL.}
\examples{
setGlobalUrl(getGlobalUrl())
}
